requirements/common.txt
dj-database-url==0.4.2
django-configurations==2.0.0
django-debug-toolbar==1.8
django-dotenv==1.4.1
django-extensions==1.7.9
Django==1.11.1
psycopg2==2.7.1
mysqlclient==1.3.10
Markdown==2.6.8